requirements.txt
langchain
langchain-community
langchain-groq
python-dotenv
streamlit
groq
tiktoken
requests
numpy
scikit-learn
sentence-transformers
